requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
SQLAlchemy==2.0.4
psycopg2-binary==2.9.6
alembic==1.10.2
pika==1.3.1
pytest==7.3.1
pytest-cov==4.0.0
bcrypt==4.0.1
pyjwt==2.8.0
alembic==1.10.2